backend/pyproject.toml
[project]
name = "openconstructionerp"
version = "0.8.0"
description = "The #1 open-source platform for construction cost estimation — BOQ, 4D/5D, AI, CAD/BIM takeoff"
license = "AGPL-3.0-or-later"
requires-python = ">=3.12"
authors = [
    { name = "Artem Boiko", email = "info@datadrivenconstruction.io" }
]
maintainers = [
    { name = "Data Driven Construction", email = "info@datadrivenconstruction.io" }
]
readme = "README.md"
keywords = [
    "construction", "cost-estimation", "boq", "bill-of-quantities",
    "quantity-surveying", "erp", "bim", "cad", "revit", "ifc",
    "din276", "nrm", "masterformat", "gaeb", "cwicr",
    "fastapi", "react", "open-source",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial",
    "Topic :: Scientific/Engineering",
    "Framework :: FastAPI",
]

dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    # Database (SQLite — works out of the box, no external DB needed)
    "sqlalchemy[asyncio]>=2.0.36",
    "aiosqlite>=0.20.0",
    "alembic>=1.14.0",
    # Validation & serialization
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "email-validator>=2.1.0",
    # Auth
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    # Export & Import
    "openpyxl>=3.1.0",
    "reportlab>=4.0.0",
    "pdfplumber>=0.11.0",
    # Utilities
    "httpx>=0.28.0",
    "orjson>=3.10.0",
    "python-dateutil>=2.9.0",
    "structlog>=24.4.0",
    "tenacity>=9.0.0",
]

[project.scripts]
openestimate = "app.cli:main"
openconstructionerp = "app.cli:main"

[project.urls]
Homepage = "https://datadrivenconstruction.io/erp"
Documentation = "https://openconstructionerp.com/docs"
Repository = "https://github.com/datadrivenconstruction/OpenConstructionERP"
Issues = "https://github.com/datadrivenconstruction/OpenConstructionERP/issues"
Changelog = "https://github.com/datadrivenconstruction/OpenConstructionERP/blob/main/CHANGELOG.md"

[project.optional-dependencies]
# PostgreSQL + Redis + S3 for production deployments
server = [
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.10",
    "celery[redis]>=5.4.0",
    "boto3>=1.35.0",
]
# Embedded vector search (LanceDB)
vector = [
    "lancedb>=0.17.0",
    "pyarrow>=18.0.0",
    "fastembed>=0.4.0",
]
# AI/LLM providers + Qdrant vector DB
ai = [
    "qdrant-client>=1.12.0",
    "sentence-transformers>=3.3.0",
    "anthropic>=0.40.0",
    "openai>=1.55.0",
]
# Computer vision for PDF takeoff
cv = [
    "paddleocr>=2.9.0",
    "ultralytics>=8.3.0",
    "opencv-python-headless>=4.10.0",
    "pymupdf>=1.25.0",
    "Pillow>=12.1.1",
]
# Everything
all = [
    "openconstructionerp[server,vector,ai,cv]",
]
# Development tools
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-xdist>=3.5.0",
    "httpx>=0.28.0",
    "factory-boy>=3.3.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
    "pyinstaller>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

# Bundle pre-built frontend into the wheel for pip install
# Build frontend first: cd frontend && npm run build
[tool.hatch.build.targets.wheel.force-include]
"../frontend/dist" = "app/_frontend_dist"

[tool.ruff]
target-version = "py312"
line-length = 120
src = ["app", "tests"]

[tool.ruff.lint]
# Pragmatic ruleset: catch real bugs (E/F/W/B), enforce import order (I)
# and Python idioms (UP/SIM/RET/PT/C4) without drowning in stylistic noise.
# Strict annotation enforcement (ANN) and unused-arg detection (ARG) are
# off by design — FastAPI's DI heavily uses unused-arg parameters and the
# project pre-dates ruff strict mode, so retroactively annotating ~1.8k
# call sites would be churn for no real safety win.
select = [
    "E", "F", "W",     # pyflakes + pycodestyle
    "I",                # isort
    "N",                # pep8-naming
    "UP",               # pyupgrade
    "B",                # bugbear
    "A",                # builtins
    "C4",               # comprehensions
    "PT",               # pytest
    "RET",              # return
    "SIM",              # simplify
]
ignore = [
    "COM812",           # trailing comma (conflicts with formatter)
    "B008",             # function calls in defaults — required by FastAPI Depends()
    "N818",             # exception name should end with Error — historical names kept
    "B904",             # raise from — not always meaningful
    "SIM105",           # contextlib.suppress — try/except is clearer for one liners
    "SIM108",           # ternary if-else — readability call, not always cleaner
    "SIM117",           # nested with — sometimes clearer separated
    "RET504",           # unnecessary assignment before return — readability call
    "RET505",           # unnecessary else after return — readability call
    "UP007",            # X | Y union syntax — keep Optional[X] readable for FastAPI
    "UP035",            # deprecated typing imports — touches too many files
    "UP038",            # non-pep604 isinstance — `isinstance(x, (int, str))` reads fine
    "E402",             # module-level imports not at top — some files set env vars first
    "E501",             # line-too-long — formatter handles wrapping where it matters
    "E701",             # multiple statements on one line — common in stub clauses
    "E702",             # multiple statements on one line (semicolon) — same
    "N806",             # non-lowercase variable in function — common for matrix/Pydantic locals
    "A002",             # builtin shadowing in args — `id`, `type`, `format` are clearest names
    "B007",             # unused loop control var — sometimes intentional for clarity
    "B017",             # assert raises Exception — fine for tests verifying any failure
    "B023",             # function uses loop variable — sometimes intentional in tests
    "PT011",            # pytest.raises with broad Exception — fine for edge tests
    "PT028",            # pytest parameter with default — fine for fixtures
    "F841",             # unused variable — IDEs catch this; sometimes intentional in tests
    "C416",             # unnecessary list/set comprehension — readability call
    "C401",             # unnecessary generator — readability call
    "SIM102",           # collapsible if — sometimes nested is clearer
    "F811",             # redefined while unused — common in test fixtures
    "RET503",           # implicit return — sometimes intentional for None
    "UP022",            # replace stdout/stderr — old pattern, harmless
    "E722",             # bare except — sometimes deliberate (silencing all errors)
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
]
markers = [
    "unit: Unit tests (no database)",
    "integration: Integration tests (requires database)",
    "slow: Slow tests (>5s)",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.coverage.run]
source = ["app"]
omit = ["app/migrations/*", "tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.",
]
